Got the assessment working, with tests.
- Had to fix some logic when calculating interest to prevent overflows. This meant just
handling the case when someone might have a "lastUpdatedBlock" that is later than their
nextDueBlock. Generally this shouldn't happen anyway, because we should be assessing you
as close to your actual due time as possible. But in the case where it does happen, we now
deal with that gracefully.
- Added a buidlerEVM config to allow for unlimited contract size. Reason is that we
really want to use buidlers console.sol for console logs during development, however before this change,
it was causing our contracts to not deploy because they were "too big", which really was caused
by the huge console.sol file (1500 lines and all internal functions, which means it gets added into the bytecode).
This config change fixes that, but shouldn't actually affect us in prod, because we'd never include console.sol in prod.
- Removed access control on sending money to the pool, because we now need the creditLine to send
to the pool as well as the credit desk. The creditLine sends during assessment. We determined this
to not be a security issue at all, because you're just sending money to the pool, so it's fine if it's public.