Take a cut for the protocol reserve.
- Added tests for this, and a new config var.
- Also added a config var for the withdraw fee, but will add tests for that in the next commit.
- Note we're doing this as a "denominator", rather than a percentage, because Solidity
doesn't support fixed point or decimal math. We could pull in a decimal library, but it's kind of overkill
for our immediate needs, since we really just need to divide by 10 and by 200 (10%, and 0.5% respectively).
- It's worth noting that dividing by those numbers could result in the fee being zero, but that would only happen
if the amount being paid back in interest is less than 10 atomic units of USDC, which is two orders of magnitude less than a penny.
- Even in this case, there is no bug. It just means that the protocol wouldn't take a fee. It would all go to the pool.
- There is also no exploit here, since it doesn't cost the borrower any less. It just shifts where the amounts go.