[Feature] - Add the transactions page (#120)
* add the transactions page

* fix mismatch with verb vs. noun tx type labels

* add connectionNotice to transactions page

* Another refactor of forms, after finding a display bug in them.

- Found a really nasty display bug in our forms that took Sanjay and I
well over an hour to find the cause of. It had to do with the timing and ordering
of re-renders, and was not at all intuitive. This led to us wanting to refactor
the interactions between components to be simpler, which is what this commit does.
- I completely deleted the useTXLoading hook, and replaced it's main functionality with
a new useSendFromUser hook, which does all the tracking of actions. The loading button itself
now handles the setIsPending, which makes more sense anyway. And this also allowed
for the loading button to become much simpler, as it no longer has to take actionComplete,
or sendFromUser or all this other more specific stuff.
- I think overall this is a much better/simpler set of componenets that are more re-usable.

* Some further cleanup

- Now the USDC Unlock form still works.
- Removed stray console logs.
- Now the sendFromUser hook automatically uses user.address, instead of
making the caller pass it in.

Co-authored-by: Blake West <blake@goldfinch.finance>
Co-authored-by: Blake <bwest87@gmail.com>