[Feature] - Borrower contract (#155)
* Started our Borrower contract.

- Added a new Borrower.sol file.
- The Borrower contract represents a bit of an architectural moment, where we
now represent the "core" protocol, and the "periphery". The Borrower is on the
"periphery". The reasoning is that the core protocol is agnostic to the fact that
we are using the Borrower contract at all. Goldfinch will provide an "official" borrower
contract for convenience, but in theory anyone else could make their own and that's fine.
- We therefore moved all the "core" stuff into it's own folder, and the "periphery" stuff also
in it's own folder.
- Added two initial passing tests. Was able to refactor out the "addressToSendTo" feature of Drawdown
from the core protocol. This type of thing is a key win of this architecture.
- More work to be done, but this is a good start.

* Got all tests passing for "drawdown" on the Borrower contract.

- Also got all other tests passing too, so whole thing is now clean.
- Added a few tests around ensuring the drawdown is locked down to only admin
of the Borrower contract.

* Added the ability to pay through the Borrower contract.

- This is a simple pass through of paying through the Borrower contract. It does
essentially no additional logic, except first transferring the funds to itself, and then
paying, so that the core protocol can assume the borrower has the funds directly.

* Switched the order of the drawdown parameters to align with `pay`.

- Now the cl address comes first, which aligns with pay.
- This also aligns with other methods such as `distributeLosses`, `createCreditLine`,
`assessCreditLine` and others, where the address always comes first. I think this also
intuitively makes more sense.

* A bit of cleanup from self review.

- Cleaned up some docs.
- Removed some stray console imports.
- Removed unnecessary _ from param names.

* Fixed up some things from review comments.

Co-authored-by: Blake West <blake@goldfinch.finance>