[Feature] - Multiple credit lines (#166)
* Added ability to switch between credit lines.

- We will need the ability to switch once our borrowers have multiple credit lines,
which we anticipate happening within the month.
- This just sets up a simple toggle between whatever credit lines you have.
- This still desperately needs CSS work, but functionally is quite close.

* add credit line dropdown styling

* Added the ability to select from multiple credit lines.

- You can now have N credit lines that you can choose from in the drop down.
- Each one will show it's limit. We could show the balance too, but I didn't
want to have to pull in all the credit line data to show that properly. And
I also didn't want to refactor that just yet. I think limit will suffice
for our immediate needs of just Luis, and we can update later if we need to.
- Fixed a small thing in our "useCloseOnClickOrEsc" hook. We now check
if the node exists first before caling contains on it. This is because I had to
use this functionality on some HTML that can conditionally be not present,
meaning the node no longer exists, and so the function call failed.

* Make changing credit line also close the CL dropdown.

Co-authored-by: Blake West <blake@goldfinch.finance>
Co-authored-by: mikesall <michael.s.sall@gmail.com>