Swapped payments frontend (#195)
* Fund protocol_owner and test user with ERC20s

* Generalize ERC20

- Add a class to hold ERC20 data (like contract and ticker)
- Add ERC20 classes for USDC, USDT, BUSD
- Refactor app state to use USDC class

* Generalize UnlockUSDCForm to UnlockERC20Form

* Add currency selector to PaymentForm

- The currency selector shows an unlock form if the currency does not have an
  allowance greater than the transaction limit
- More functionality to come later

* Show 1Inch quotes

* Fix pre-defined amount estimates

Before, we were using a 1Inch quote from USDC to the other ERC20 token. This always
produced a lower value than we wanted due to slippage. Now we get a quote and
estimate a reasonable starting value using the quoted spread + some padding.

* Fix unlock form margin

* Remove number type

It prevents form submission in firefox due to validating an integer by default. We
could change the validation to support decimal places, but we already do our own validation anyway.

* Do the swapped payment

* Make minimum configurable

* Rename top-level erc20 state to usdc

It's actually always USDC and now that we support multiple ERC20s, the naming is confusing.

* Add loading states and use them in some places

* Use whales instead of 1inch to fund

Loading 1inch state was too slow to be usable

* Expose getter for address

* Remove unused local contract names

* Refactor into method, use ethers instead of web3