[Chore] Goldfinch Config extensibility (#210)
* Make GoldfinchConfig extensible with new numbers/addresses without upgrade.

Currently, adding any new addresses or numbers required updating the ConfigOptions,
which implicitly required upgrading/changing the GoldfinchConfig. Because we took
away the proxy aspect of GoldfinchConfig (to reduce gas costs), we can no longer
easily upgrade GoldfinchConfig. Hence, we are removing the dependency on ConfigOptions
when setting new addresses or numbers.

* Remove link calls between ConfigOptions and GoldfinchConfig.

- We don't actually need to link these anymore, since the only
usage of ConfigOptions with GoldfinchConfig is a constant, and the compiler
automatically just puts the library in the bytecode directly.

* Fixing other tests related to the GoldfinchConfig changes.

- Also remove other places where we're trying to link ConfigOptions
with GoldfinchConfig, which is no longer necessary.