[Feature] Implement SeniorFund pool-like methods + invest (#236)
* SeniorFund with Pool-like methods

Partial implementation of SeniorFund with Pool-like functionality e.g. deposit,
withdraw, sweep to/from compound

* Use separate SeniorFundFidu

This is a short-term way to keep Pool tests passing. When we deploy, we may want to repurpose Fidu
for this without deploying a separate token, or we may want to migrate by minting
SeniorFundFidu and burning Fidu.

* Implement FixedLeverageRatioStrategy

- Implement the strategy
- Deploy it with a leverageRatio of 4
- Add unit tests with a TestTranchedPool

* Implement SeniorFund.invest

* Stub out redeem and writedown tests

* Cleanup: Re-order methods and add natspecs

* Fix after rebasing

* Use enum

* Don't bubble up overflow errors

* Implement validPool

* Remove SeniorFundFidu

* Clean-up

* Revert "Remove SeniorFundFidu"

This reverts commit dbb27aa1b8f5aef918494bd58dbb6538bbee1f3d.