[Feature] - Create pools pt 2 (#259)
* Refactored TranchedPool creation.

- CreditLines are now created from within the initializer of the TranchedPool
- We fully bypass the CreditDesk now. No storing anything in mappings.
- Underwriters have been removed as a variable from CreditLines
- Fixed up the minting test so that it still works.

* Refactor to use a mapping for pool validity.

- I tried to pull out the logic address from the minimal proxy, and do that.
However, it would have required upgrading to Solidity 0.8.0, which honestly wouldn't be
too bad, but still seems to add unnecessary complexity.
- Perhaps more importantly, upon further reflection, it wasn't clear what benefit we got
by doing it the "pure" create2 way, because it wouldn't be cheaper. We'd still have to
do a storage read for the address. So... then we realized let's just do a mapping which
is simpler and doesn't require an upgrade.

* Fix a bunch of tests, resulting from the changes in how CreditLines and Pools get created.

- Pools are no longer created through the CreditDesk.
- Also Underwriters as a concept are gone from the new CreditLines.

* Rename markPoolCreated -> onPoolCreated