Revert "Turn on noImplicitOverride."
This reverts commit a9f5da88a1fc74c2b67684be6eda2244801489c9.

Babel does not understand how to transpile TS that contains the
`override` keyword (it raises a syntax error), which was introduced
only in TS 4.3 (i.e. the latest minor version of TS). Even after
upgrading react-scripts to v4, which entails upgrading babel-jest
to v26, Babel still can't transpile the `override` keyword. It's
possible that v27 of babel-jest contains a fix, but we aren't
able to upgrade to v27 of babel-jest yet because react-scripts v4
complains about that and only wants v26.