Make localhost network explicit in hardhat commands.
This seems necessary by virtue of the fact that for the murmuration deploy,
we run `hardhat deploy` as a separate command from `hardhat node --no-deploy`.
When using `hardhat node` to deploy (i.e. without the `--no-deploy` flag),
hardhat gives the deployment the name `localhost` by default. But we observed
the hardhat deployment deployed on App Engine to have the name `hardhat`.
That name must be coming from the default behavior of the `hardhat deploy`
command; hopefully providing the `--network localhost` argument, which
is a global option to hardhat commands, fixes that and makes the
deployment name `localhost`.