[Fix] withdrawSenior deducts too much fidu if compound is swept (#444)
Previously it calculated the number of shares to deduct from the position before
withdrawing from the SeniorPool. But, if the SeniorPool sweeps from compound, the
share price will increase, so fewer fidu are required to withdraw the same amount of
usdc. It now correctly accounts for this case by looking at the actual fidu balance
burned as a result of withdrawal.