[Fix/Refactor] Connection notice not showing unlock (#552)
* [Fix/Refactor] Connection notice not showing unlock

- Refactor ConnectionNotice to make match + render logic really explicit. In
particular, cases are now exclusively matched, in-order. There can be no fall-through to
other cases when a case is matched, making this equivalent using else-ifs or always
returning from inside the condition block. New conditions can be added by creating a
new entry in the `strategies` constant.
- Add loading state when KYC status is being fetched and error state when KYC request
fails. Makes for a more clear user experience.
- Fix USDC unlock not showing when KYC is successful. This was happening due to the
nesting inside the requireKYC condition. So, the condition was matched but rendered
nothing because KYC was successful. I did this to prevent flickering while KYC was
loading, but I missed this bug. KYC loading is now taken care of separately

* Component tests

* See all tests pass with existing implementation

* Add back refactored logic