[fix] Autotask build (#584)
* [fix] Autotask build

Bundling as commonjs caused the output to use the peer dependency tslib. I couldn't
figure out a way to get tslib to be included with the bundle, even after installing
it as a dependency. Using es6 as the module type causes tslib not to be required in
the generated code, bypassing the issue. The defender-autotask-examples repo uses es6
modules, so this should theoretically work fine: https://github.com/OpenZeppelin/defender-autotask-examples/blob/master/rollup/tsconfig.json#L6

* Fix test build error