[fix] BackerRewards incorrect for backers who withdrew before TranchedPool is locked (#407)
* [fix] BackerRewards incorrect for backers who withdrew before TranchedPool is locked

Decrement PoolToken principalAmount rather than increment principalRedeem if the
TranchedPool.withdraw is happening before the Tranche is locked. This aligns with
BackerRewards' assumption that it should grant rewards based on a PoolTokens
principalAmount without regard to principalRedeemed. This also makes intuitive sense:
we separate the concept of "redeeming" after a pool is locked from the concept of
moving capital freely in/out before a pool is locked.

* Fix mainnet-forking tests

- Need to upgrade PoolTokens so that it has the extra method

* Fix Zapper expectation

* Fix mainnet forking test

* Add tests for new PoolTokens method

* Test that principalRedeemed is incremented on withdraw after lock

* Test the fix

* Review comments

* Fix test (goes through new codepath with different revert message)