Add destroyUser cloud function for burnUID script (#679)
### What happened?
I added a new cloud function called _destroyUser_ to help with the burnUID script. The function is gated behind signature verification and a signer whitelist. I call this new auth "signatureWithAllowList". The allow list for _destroyUser_ is hardcoded to be the trusted Unique Identity Signer.

After verifying the signature and signer, the function will delete the user from the `users` store and write to the `destroyedUsers` store. If it's the first deletion for the address then it will create a new document in `destroyedUsers`. On subsequent deletions it'll append to the existing one (so we can keep track of multiple burns).

### Why?
This cloud function is one part of the burnUID script. Full PR here: https://github.com/warbler-labs/mono/pull/431. Also see that PR description for a diagram of how the cloud function fits into the big picture.

### How has this been tested?
- [x] Unit tested
- [x] Tested locally when I tested the entire script in the other PR.

**Checklist**
- [x] I have reviewed this code myself
- [ ] Screenshots of the key change are present (if this is a front-end PR)