add to zapper natspec (#689)
### What happened?
Add a comment to _zapStakeToTranchedPool_ about needing to ERC721.approve the Zapper address before calling the function.

### Why?
Calling ERC721.approve(zapperAddress,...) is a necessary condition for zapStakeToTranchedPool to succeed, but it's not immediately clear from the existing NatSpec documentation or the code. @pugbyte and I have had to chase down the issue on separate occasions and we could have saved ourselves time if it was documented on the function. It will also save us time in the future (and the time of future third party integrators).

### How has this been tested?
No tests necesary

**Checklist**
- [x] I have reviewed this code myself
- [ ] Screenshots of the key change are present (if this is a front-end PR)