Add zapMultiple, unzapMultiple, and claimMultiple to Zapper (#673)
### What happened?
Added three public functions to the Zapper contract: _zapMultiple_, _unzapMultiple_, and _claimMultiple_ which serve as the batch analogues of _zapStakeToTranchedPool_, _unzapToStakingRewards_, and _claimTranchedPoolZap_ respectively. The new functions enable zapping, unzapping, and claiming multiple tokens in a single transaction.

**Note**: CI is failing due to the oversized **StakingRewards** contract but @wbj-goldfinch is working on a fix for that.

### Why?
Having to zap, unzap, and claim individual staking rewards or pool tokens is a bad user experience. The user has to sign a transaction for every action. Presently they see different "positions" in a drop down for them to zap and claim. This should be abstracted away from the user like we do for TranchedPool _withdrawMultiple_.

Below is a screenshot of the current UX, which we can improve once _zapMultiple_ is live.
<img width="967" alt="Screen Shot 2022-06-06 at 10 14 26 PM" src="https://user-images.githubusercontent.com/19829228/172301067-9493564f-c9f0-4332-b63f-f3dba6904b65.png">

### How has this been tested?
#### Unit tests
- [x] zapMultiple
- [x] unzapMultiple
- [x] claimMultiple

**Checklist**
- [x] I have reviewed this code myself
- [x] Screenshots of the key change are present (if this is a front-end PR)