Signature verification with arbitrary signatures (#752)
### What happened?
Remove the hardcoded `Sign in with Goldfinch: ${blockNum}` plaintext server side and instead pass the plaintext in the `x-goldfinch-signature-plaintext` header. This will allow callers to tell _verifySignature_ what plaintext to verify the signature against.

In addition to making the signature's plaintext configurable, the signature's max age is also configurable. _destroyUser_ is a more sensitive operation than _kycStatus_ so it merits a quicker expiry.

### Why?
@ihinsdale raised the valid concern that using `Sign in with Goldfinch: ${blockNum}` for the signature on _destroyUser_ is a security vulnerability because this signature isn't scoped to a particular address and UID type. An intercepted signature could be used to call _destroyUser_ to delete any and all accounts from our database.

To scope the signature to an address and UID type we can encode the address and UID type in the plaintext that gets signed. _verifySignature_ will verify the signature and then _destroyUser_ can parse the address and UID type out of the plaintext message. Under this new scheme a replay on _destroyUser_ would only be valid on the same address and UID pair.

### Backwards compatibility
The client currently calls /kycStatus but doesn't pass it the `x-goldfinch-signature-plaintext` header. If the new cloud functions are deployed before the client then this will cause a spike in errors because the old client will be sending requests to /kycStatus without the x-goldfinch-signature-plaintext header.

To avoid this we added a temporary `fallbackOnMissingPlaintext` flag to verifySignature so that when the `x-goldfinch-signature-plaintext` header is missing, /kycStatus can fallback to using the `Sign in to Goldfinch...` plaintext. The fallback will be removed in a followup PR after the client (and client2) have been updated to use the x-goldfinch-signature-plaintext header.

#### Other PRs
Use x-goldfinch-signature-plaintext header in the new client: https://github.com/warbler-labs/mono/pull/761

### How has this been tested?
- [x] Unit tests

**Checklist**
- [x] I have reviewed this code myself
- [ ] Screenshots of the key change are present (if this is a front-end PR)